refactor config readin
