+++ src/gui/mod.rs
@@ -23,13 +23,13 @@ use self::playback::play_excerpt;
 use self::playback::PlaybackThreadHandle;
 use self::plot::ExcerptPlot;
 use crate::audio_time::AudioTime;
+use crate::config::Opts;
 use crate::cut::CutInfo;
 use crate::excerpt_collection::ExcerptCollection;
 use crate::gui::session_manager::SessionIdentifier;
 use crate::gui::session_manager::SessionManager;
 use crate::song::format_title;
 use crate::song::Song;
-use crate::Opts;
 
 #[derive(PartialEq, Eq, Copy, Clone)]
 struct SongIdentifier {
